+++ packages/auth/src/features/verifyCode/service.ts
@@ -12,10 +12,11 @@ import { redis } from '@auth/database'
 import { JWT_ISSUER } from '@auth/constants'
 import { env } from '@auth/environment'
 import * as crypto from 'crypto'
-import { IUserName, createToken } from '@auth/features/authenticate/service'
+import { createToken } from '@auth/features/authenticate/service'
 import {
   triggerUserEventNotification,
-  personNameFromV1ToV2
+  personNameFromV1ToV2,
+  IUserName
 } from '@opencrvs/commons'
 
 interface ICodeDetails {
