+++ packages/client/src/v2-events/features/users/cache.ts
@@ -11,11 +11,7 @@
 
 import { flatten, uniq } from 'lodash'
 import { EventDocument, EventIndex } from '@opencrvs/commons/client'
-import {
-  queryClient,
-  trpcClient,
-  trpcOptionsProxy
-} from '@client/v2-events/trpc'
+import { queryClient, trpcOptionsProxy } from '@client/v2-events/trpc'
 import { findUserIdsFromDocument, findUserIdsFromIndex } from './utils'
 
 async function cacheUsers(userIds: string[]) {
